+++ tapo/src/responses/device_info_result/power_strip.rs
@@ -3,8 +3,9 @@ use serde::{Deserialize, Serialize};
 use crate::error::Error;
 use crate::responses::{decode_value, DecodableResultExt, TapoResponseExt};
 
-/// Device info of Tapo H100. Superset of [`crate::responses::DeviceInfoGenericResult`].
+/// Device info of Tapo P300 and P304. Superset of [`crate::responses::DeviceInfoGenericResult`].
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
 pub struct DeviceInfoPowerStripResult {
     //
@@ -32,6 +33,18 @@ pub struct DeviceInfoPowerStripResult {
     pub latitude: Option<i64>,
     pub longitude: Option<i64>,
     pub time_diff: Option<i64>,
+}
+
+#[cfg(feature = "python")]
+#[pyo3::pymethods]
+impl DeviceInfoPowerStripResult {
+    /// Gets all the properties of this result as a dictionary.
+    pub fn to_dict(&self, py: pyo3::Python) -> pyo3::PyResult<pyo3::Py<pyo3::types::PyDict>> {
+        let value = serde_json::to_value(self)
+            .map_err(|e| pyo3::exceptions::PyException::new_err(e.to_string()))?;
+
+        crate::python::serde_object_to_py_dict(py, &value)
+    }
 }
 
 impl TapoResponseExt for DeviceInfoPowerStripResult {}