+++ tapo/src/responses/energy_usage_result.rs
@@ -2,7 +2,7 @@ use chrono::NaiveDateTime;
 use serde::{Deserialize, Serialize};
 
 use crate::responses::TapoResponseExt;
-use crate::tapo_date_format::der_tapo_datetime_format;
+use crate::utils::der_tapo_datetime_format;
 
 /// Contains local time, current power and the energy usage and runtime for today and for the current month.
 #[derive(Debug, Serialize, Deserialize)]
@@ -11,44 +11,16 @@ pub struct EnergyUsageResult {
     /// Local time of the device.
     #[serde(deserialize_with = "der_tapo_datetime_format")]
     pub local_time: NaiveDateTime,
-    /// Current power in milliwatts (mW).
+    /// Current power in Milliwatts (mW).
     pub current_power: u64,
     /// Today runtime in minutes.
     pub today_runtime: u64,
-    /// Today energy usage in watts (W).
+    /// Today energy usage in Watt Hours (Wh).
     pub today_energy: u64,
     /// Current month runtime in minutes.
     pub month_runtime: u64,
-    /// Current month energy usage in watts (W).
+    /// Current month energy usage in Watt Hours (Wh).
     pub month_energy: u64,
-    /// Hourly energy usage for the past 24 hours in watts (W).
-    #[deprecated(
-        since = "0.4.0",
-        note = "P110 firmware v1.1.6 no longer returns this field. Use `tapo::PlugEnergyMonitoringHandler::get_energy_data` instead."
-    )]
-    #[cfg(not(feature = "python"))]
-    pub past24h: Option<Vec<u64>>,
-    /// Hourly energy usage by day for the past 7 days in watts (W).
-    #[deprecated(
-        since = "0.4.0",
-        note = "P110 firmware v1.1.6 no longer returns this field. Use `tapo::PlugEnergyMonitoringHandler::get_energy_data` instead."
-    )]
-    #[cfg(not(feature = "python"))]
-    pub past7d: Option<Vec<Vec<u64>>>,
-    /// Daily energy usage for the past 30 days in watts (W).
-    #[deprecated(
-        since = "0.4.0",
-        note = "P110 firmware v1.1.6 no longer returns this field. Use `tapo::PlugEnergyMonitoringHandler::get_energy_data` instead."
-    )]
-    #[cfg(not(feature = "python"))]
-    pub past30d: Option<Vec<u64>>,
-    /// Monthly energy usage for the past year in watts (W).
-    #[deprecated(
-        since = "0.4.0",
-        note = "P110 firmware v1.1.6 no longer returns this field. Use `tapo::PlugEnergyMonitoringHandler::get_energy_data` instead."
-    )]
-    #[cfg(not(feature = "python"))]
-    pub past1y: Option<Vec<u64>>,
 }
 impl TapoResponseExt for EnergyUsageResult {}